client/requirements.txt
pyzmq
protobuf
numpy
matplotlib
scipy
py3gpp